app_web/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.3
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
